BuildConfig: added Project class
